chore: moves locale types to assistTypes.ts
